made cheese and point bonuses. but need to improve upon cheese spawn location, which right now is pretty crude
