+++ Assembly-CSharp.csproj
@@ -42,14 +42,15 @@
      <Compile Include="Assets\Scripts\PigeonSpawn.cs" />
      <Compile Include="Assets\Scripts\CatMovement.cs" />
      <Compile Include="Assets\Scripts\BGMusic.cs" />
+     <Compile Include="Assets\Scripts\backgroundScale.cs" />
      <Compile Include="Assets\Scripts\BuildingSpawn.cs" />
-     <Compile Include="Assets\RestartGame.cs" />
+     <Compile Include="Assets\Scripts\RestartGame.cs" />
      <Compile Include="Assets\Scripts\PigeonMovement.cs" />
      <Compile Include="Assets\Scripts\buildingFall.cs" />
      <Compile Include="Assets\Scripts\streetParallax.cs" />
      <Compile Include="Assets\Scripts\streetView.cs" />
      <Compile Include="Assets\Scripts\CanvasFade.cs" />
-     <Compile Include="Assets\Scripts\backgroundScale.cs" />
+     <Compile Include="Assets\Scripts\cheeseSpawn.cs" />
     <Reference Include="UnityEngine">
         <HintPath>/Applications/Unity/Hub/Editor/2020.1.2f1/Unity.app/Contents/Managed/UnityEngine/UnityEngine.dll</HintPath>
     </Reference>
