+++ Assets/Animations/cheeseAnim.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 755d66fc0586d48a1bb86587c59a6c60
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
