+++ Assets/Animations/cheeseSparkle.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bd253167fee4f4e30bc8d507568347a8
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
