+++ Assets/Animations/cheeseSparkleAnim.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3d8387794c187444bb6fffccd6da4346
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
