+++ Assets/Images/New Animation.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b046c64b261f64d2f91af444446eadf3
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
