--- Assets/RestartGame.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6703c9492ceef415e9a529b3c5380a87
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
