+++ Assets/Scripts/CanvasFade.cs
@@ -46,8 +46,8 @@ public class CanvasFade : MonoBehaviour
             StartCoroutine(FadeText(secondMessage, 0f, 1f, 0.5f));
             yield return new WaitForSeconds(3f);
             StartCoroutine(FadeText(secondMessage, 1f, 0f, 0.5f));
-            pigeonSpawn.GetComponent<PigeonSpawn>().canSpawn = true;
         }
+        pigeonSpawn.GetComponent<PigeonSpawn>().canSpawn = true;
     }
 
     IEnumerator FadeText(CanvasGroup text, float startAlpha, float endAlpha, float duration)
