+++ Assets/Scripts/cheeseSpawn.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f8dbd932a7aaa4be4b2cdad33263ee6d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
