+++ ProjectSettings/TagManager.asset
@@ -7,6 +7,7 @@ TagManager:
   - Building
   - Street
   - Pigeon
+  - Cheese
   layers:
   - Default
   - TransparentFX
